# api/openapi.yaml
openapi: 3.1.0
info:
  title: BotHub API
  version: 0.1.0
  description: |
    Contract between the Go dashboard and the PHP API.

    Rules:
    - The API is the only way to change configuration. The dashboard never touches the DB.
    - Every route except `/api/health`, `GET /api/v1/setup`, `POST /api/v1/setup` and
      `POST /api/v1/auth/login` needs a valid session. One central middleware checks this.
    - Every changing request (POST, PUT, PATCH, DELETE) needs the header `X-CSRF-Token`.
    - Errors never contain UI text. They contain an i18n key that the dashboard translates.
    - Bot tokens are write-only. The API never returns them.
    - Actions that need Discord return `202` with a job. Poll `GET /api/v1/jobs/{jobId}`.
servers:
  - url: /
security:
  - session: []
...
paths:
...
  /api/v1/setup:
    get:
      operationId: getSetupState
      tags: [auth]
      summary: Whether the first-start setup is still required
      security: []
      responses:
        "200":
          description: Setup state
          content:
            application/json:
              schema:
                type: object
                required: [required]
                properties:
                  required: { type: boolean }
    post:
      operationId: runSetup
      tags: [auth]
      summary: Create the single admin user (only while setup is required)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetupRequest" }
      responses:
        "201":
          description: Admin created and logged in
          headers:
            Set-Cookie: { $ref: "#/components/headers/SessionCookie" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Me" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
...
  /api/v1/settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Instance settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Settings" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    patch:
      operationId: updateSettings
      tags: [settings]
      summary: Change instance settings
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Settings" }
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Settings" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /api/v1/bots:
    get:
      operationId: listBots
      tags: [bots]
      summary: List all bots
      responses:
        "200":
          description: Bots
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Bot" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      operationId: createBot
      tags: [bots]
      summary: Add a bot
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BotCreate" }
      responses:
        "201":
          description: Bot created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Bot" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /api/v1/bots/{botId}:
    parameters:
      - $ref: "#/components/parameters/BotId"
    get:
      operationId: getBot
      tags: [bots]
      summary: Get one bot
      responses:
        "200":
          description: Bot
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Bot" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateBot
      tags: [bots]
      summary: Change name, token or autostart
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BotUpdate" }
      responses:
        "200":
          description: Bot
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Bot" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationError" }
    delete:
      operationId: deleteBot
      tags: [bots]
      summary: Delete a bot and all its data
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/bots/{botId}/start:
    post:
      operationId: startBot
      tags: [bots]
      summary: Start a bot
      parameters:
        - $ref: "#/components/parameters/BotId"
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "202": { $ref: "#/components/responses/JobAccepted" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/bots/{botId}/stop:
    post:
      operationId: stopBot
      tags: [bots]
      summary: Stop a bot
      parameters:
        - $ref: "#/components/parameters/BotId"
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "202": { $ref: "#/components/responses/JobAccepted" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/bots/{botId}/guilds:
    get:
      operationId: listGuilds
      tags: [guilds]
      summary: Guilds the bot is a member of (from the bot's cache)
      parameters:
        - $ref: "#/components/parameters/BotId"
      responses:
        "200":
          description: Guilds
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Guild" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/bots/{botId}/guilds/{guildId}:
    delete:
      tags: [guilds]
      operationId: leaveGuild
      summary: Make the bot leave a guild
      parameters:
        - $ref: "#/components/parameters/BotId"
        - $ref: "#/components/parameters/GuildId"
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "202": { $ref: "#/components/responses/JobAccepted" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: bothub_session
  headers:
    SessionCookie:
      description: "`bothub_session`, HttpOnly, SameSite=Strict, Path=/"
      schema: { type: string }
  parameters:
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: Token from `GET /api/v1/auth/me`
      schema: { type: string }
    BotId:
      name: botId
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
    GuildId:
      name: guildId
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Snowflake" }
...
  responses:
    Unauthorized:
      description: No valid session (`error.auth.required`, `error.auth.invalid_credentials`)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: CSRF token missing or wrong (`error.csrf.invalid`)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found (`error.<resource>.not_found`)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: State does not allow this action
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ValidationError:
      description: Request body invalid
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    JobAccepted:
      description: Job queued for the bot
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Job" }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [key]
          properties:
            key:
              type: string
              description: i18n key, e.g. `error.bot.not_found`
              examples: [error.bot.not_found]
            params:
              type: object
              description: Values for placeholders in the translated text
              additionalProperties: true
            fields:
              type: object
              description: Per-field i18n keys for validation errors
              additionalProperties: { type: string }
    Snowflake:
      type: string
      pattern: "^[0-9]{17,20}$"
    Locale:
      type: string
      pattern: "^[a-z]{2}(-[A-Z]{2})?$"
      default: en
    SetupRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string, minLength: 3, maxLength: 32 }
        password: { type: string, minLength: 12, maxLength: 256 }
        locale: { $ref: "#/components/schemas/Locale" }
...
    Me:
      type: object
      required: [username, locale, theme, csrfToken]
      properties:
        username: { type: string }
        locale: { $ref: "#/components/schemas/Locale" }
        theme:
          type: string
          enum: [system, light, dark]
        csrfToken: { type: string }
    Settings:
      type: object
      properties:
        locale: { $ref: "#/components/schemas/Locale" }
        theme:
          type: string
          enum: [system, light, dark]
    BotStatus:
      type: string
      enum: [stopped, starting, running, stopping, error]
    Bot:
      type: object
      required: [id, name, status, tokenSet, autostart, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        applicationId: { oneOf: [{ $ref: "#/components/schemas/Snowflake" }, { type: "null" }] }
        avatarUrl: { type: [string, "null"], format: uri }
        status: { $ref: "#/components/schemas/BotStatus" }
        statusErrorKey:
          type: [string, "null"]
          description: i18n key when status is `error`
        tokenSet: { type: boolean }
        autostart: { type: boolean }
        guildCount: { type: integer }
        createdAt: { type: string, format: date-time }
        startedAt:
          type: [string, "null"]
          format: date-time
          description: When the bot went online; null unless status is `running` (uptime)
    BotCreate:
      type: object
      required: [token]
      description: |
        The API validates the token against Discord (`GET /users/@me`) and takes
        name, application ID and avatar from there. `name` overrides the Discord name.
      properties:
        name: { type: string, minLength: 1, maxLength: 64 }
        token: { type: string, writeOnly: true }
        autostart: { type: boolean, default: true }
    BotUpdate:
      type: object
      minProperties: 1
      properties:
        name: { type: string, minLength: 1, maxLength: 64 }
        token: { type: string, writeOnly: true }
        autostart: { type: boolean }
    Guild:
      type: object
      required: [id, name, memberCount]
      properties:
        id: { $ref: "#/components/schemas/Snowflake" }
        name: { type: string }
        iconUrl: { type: [string, "null"], format: uri }
        memberCount: { type: integer }
...
    Job:
      type: object
      required: [id, type, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        type:
          type: string
          enum: [bot.start, bot.stop, bot.restart, guild.leave, plugin.install, plugin.update, plugin.uninstall]
        status:
          type: string
          enum: [queued, running, done, failed]
        errorKey: { type: [string, "null"] }
        createdAt: { type: string, format: date-time }
        finishedAt: { type: [string, "null"], format: date-time }